+++ src/main.js
@@ -14,6 +14,7 @@ import recipe from "./recipe.js";
 import jobPosting from "./jobPosting.js";
 import course from "./course.js";
 import book from "./book.js";
+import techArticle from "./techArticle.js";
 
 /**
  * Structured data for the current context.
@@ -102,6 +103,8 @@ export default ({ meta, type, tags = [] }) => {
     content = course({ meta });
   } else if (type === "book") {
     content = book({ meta });
+  } else if (type === "techArticle") {
+    content = techArticle({ meta, tags });
   }
 
   return {